+++ crates/crab-protocol/src/versions/mod.rs
@@ -8,6 +8,7 @@ pub mod v1_20_1;
 pub mod v1_20_2;
 pub mod v1_20_3;
 pub mod v1_20_5;
+pub mod v1_21;
 
 /// Protocol number for Minecraft 1.20 and 1.20.1.
 pub const PROTOCOL_1_20_1: i32 = 763;
@@ -17,3 +18,5 @@ pub const PROTOCOL_1_20_2: i32 = 764;
 pub const PROTOCOL_1_20_3: i32 = 765;
 /// Java Edition 1.20.5/1.20.6.
 pub const PROTOCOL_1_20_5: i32 = 766;
+/// Java Edition 1.21/1.21.1.
+pub const PROTOCOL_1_21: i32 = 767;